sample-fastapi/pyproject.toml
[project]
name = "sample-fastapi"
version = "0.1.0"
description = "LGTM + Pyroscope test FastAPI app"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyroscope-io>=0.8.0",
    "pyroscope-otel>=0.1.0",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sample_fastapi"]
